AT-6302 have added header as vuetify recommends
