+++ src/components/StyledFields.vue
@@ -12,14 +12,14 @@
     </v-row>
     <v-row class="text-center">
       <v-col cols="12" class="mb-4">
-        <h1 class="font-weight-bold mb-3">Hello World</h1>
+        <h1 class="h1 font-weight-bold mb-3">Hello World</h1>
       </v-col>
     </v-row>
     <v-row class="text-center">
       <v-col cols="12" class="mb-4">
         <v-icon large color="success"> mdi-domain </v-icon>
         <v-icon large color="primary"> mdi-message-text </v-icon>
-        <a href="google.com" target="blank"> google.com</a>
+        <a href="google.com" target="blank" class="link-body-lg"> google.com</a>
       </v-col>
     </v-row>
     <v-row>
