+++ src/main.ts
@@ -10,17 +10,23 @@ import cssVars from 'css-vars-ponyfill'; // needed to process css vars in IE11
 
 import ATATFooter from "./components/ATATFooter.vue";
 import ATATSelect from "./components/ATATSelect.vue";
+import ATATHeader from "./components/ATATHeader.vue";
+import ATATHeaderNav from "./components/ATATHeaderNav.vue";
+import ATATSideBar from "./components/ATATSideBar.vue";
 import ATATTextField from "./components/ATATTextField.vue";
+import SecurityBanner from "./components/SecurityBanner.vue";
 import StyledFields from "./components/StyledFields.vue";
-import ATATSideBar from "./components/ATATSideBar.vue"
-
+import USWDCBanner from "./components/USWDCBanner.vue";
 
 Vue.component('atat-footer', ATATFooter);
+Vue.component('atat-header', ATATHeader);
+Vue.component('atat-header-nav', ATATHeaderNav);
 Vue.component('atat-select', ATATSelect);
 Vue.component('atat-sidebar', ATATSideBar);
 Vue.component('atat-text-field', ATATTextField);
+Vue.component('security-banner', SecurityBanner);
 Vue.component('styled-fields', StyledFields);
-
+Vue.component('usdwc-banner', USWDCBanner);
 
 Vue.config.productionTip = false;
 Vue.prototype.moment = moment;
