+++ src/sass/atat.scss
@@ -1,5 +1,11 @@
 @import "./components/v-text-field.scss";
 @import "./components/v-alerts";
 @import "./utils/typography.scss";
+@import "./utils/colors.scss";
 @import "./components/v-button";
 @import "./components/v-list-item";
+@import "./components/app-bar-top-nav";
+
+body,html{
+    font-size: 10px !important;
+}
