+++ src/sass/utils/colors.scss
@@ -1,0 +1,26 @@
+  $primary_lighter: #d9e8f6 !default;
+  $primary_light: #73b3e7 !default;
+  $primary: #005EA2 !default;
+  $primary_dark: #1a4480 !default;
+  $primary_darken: #162E51 !default;
+  $primary_vivid: #0050d8 !default;
+
+  $tertiary: #162E51 !default;
+  $success: #2E8540 !default;
+  $error: #E21C3D !default;
+
+  $warning_lighter: #faf3d1 !default;
+  $warning_light: #fee685 !default;
+  $warning: #ffbe2e !default;
+  $warning_dark: #e5a000 !default;
+  $warning_darker: #936f38 !default;
+
+  $info_darker: #2e6276 !default;
+  $info_dark: #009ec1 !default;
+  $info: #05bfe7 !default;
+  $info_light: #99deea !default;
+  $info_lighter: #e7f6f8 !default;
+
+  $hr_color: #DFE1E2 !default;
+
+  $unclass_header:  #008000 !default;
